pyproject.toml
[tool.poetry]
name = "gofish"
version = "0.1.0"
description = ""
authors = ["Pandaptable <panda@pandaptable.moe>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyAutoGUI = "^0.9.54"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
